Ability to force refresh on files
Files may be open and written to without changing properties (log4net).
A sufficiently busy system (where logging and scanning is occurring)
will cause files to have incorrect (old) attributes even if
FileSystemInfo.Refresh is called (indirectly) which is against
Microsoft's documentation on this function but not against the functions
used in the underlying implementation.  I have filed a bug report with
the Microsoft security team to investigate.